feat(CopilotChat): Improve line separation and prompt update logic, add diff view (#99)
* feat(CopilotChat): Improve line separation and prompt update logic

Now submit_code and submit_prompt properly try to find first match from the end
so you dont need to cursor over specific blocks. This is also prerequisite for
showing current diff as float or in corner somewhere.

BREAKING CHANGE: The key mapping for submitting code is now <C-d> instead
of <C-y>.

Signed-off-by: Tomas Slusny <ts6234@att.com>

* Improve the default prompts slightly and fix issue with active selection

Signed-off-by: Tomas Slusny <ts6234@att.com>

* feat: Add diff display

For now with keybinding (default K). Maybe find a way to display it in better place.

Signed-off-by: Tomas Slusny <ts6234@att.com>

* Adjust keybindings for accept_diff and show_diff

---------

Signed-off-by: Tomas Slusny <ts6234@att.com>
Co-authored-by: Tomas Slusny <ts6234@att.com>