feat: Move cursor to end of buffer before spinner start
The change ensures that the cursor is moved to the end of thebuffer before the spinner starts. This improves the user experience bymaking The change ensures that the cursor is moved to the end of thebuffer before the spinner starts. This improves the user experience bymaking sure the spinner always appears at the current end of the chat,regardless of where the cursor was previously.