feat: Add source parameter to callback function (#219)
* feat: Add source parameter to callback function

The callback function in both the 'ask' and 'config' methods now
receives the 'source' parameter. This allows users to access
information about the source of the response, such as the buffer
number and window number. This change enhances the flexibility and
usability of the callback function.

* feat: Add source parameter to callback function

The callback function in both the 'ask' and 'config' methods now
receives the 'source' parameter. This allows users to access
information about the source of the response, such as the buffer
number and window number. This change enhances the flexibility and
usability of the callback function.

* fix(CopilotChat.txt): revert change

* revert txt change