refactor: improve semantic embeddings performance
Previously the semantic embeddings were using older model and not using the
query in most optimal way. Now it is using text-embedding-3-small model for
reduced cost and better performance, embeddings are stored in 512
dimensions and the embedding process was refactored to be more
straightforward and use less API calls by including the query in the
initial embedding call.

Also slightly improves the formatting of the embedding prompt to be more
readable.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>