feat: add jump to diff feature with `gj` mapping
Adds new mapping that allows jumping to the target file and line number from
the diff view. If the target file is already open in a buffer, it will switch
to that buffer. Otherwise, it creates a new buffer with proper filetype.

This improves navigation between diff view and source files, making it easier
to review and apply suggested changes.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>