fix: make token counting optional in chat client
Currently token counting and limiting messages is enforced for all chat
providers, but this should be optional as not all chat providers support
token limiting. This change makes token counting optional and only enabled
when max_tokens and tokenizer are provided in model config.

The types in providers.lua have also been updated to reflect the optional
nature of these fields.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>