refactor: move embedding cache to context module
Move embedding cache functionality from client to context module to improve
separation of concerns and code organization. The cache is now maintained in
the context module which is more appropriate since it deals with context
management.

This change simplifies the client module by removing the embedding cache
responsibility while maintaining the same caching functionality in a more
logical location.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>