fix: functions are missing from parse_schema's return
I'll use the `file` function's schema as an example of the problem:

```lua
    schema = {
      type = 'object',
      required = { 'path' },
      properties = {
        path = {
          type = 'string',
          description = 'Path to file to include in chat context.',
          enum = function(source)
            local glob = utils.glob(source.cwd(), {
              max_count = 0,
            })
            return glob
          end,
        },
      },
    }
```

At the end of `parse_schema`, the schema is filtered:

```lua
  if schema then
    schema = filter_schema(schema, true)
  end

  return schema
```

This change leaves functions as is in the results table.
Prior to this, completing `#file:` in the picker would give me an input field to type my file path. This restores the behavior visible in the demo in the README where that file can be picked using `vim.ui.select`.